Nicolas Regnault: - fix a bug in BosonOnTorusShort for large Hilbert spaces
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@3005 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ FQHE/src/HilbertSpace/BosonOnTorusShort.h
@@ -463,7 +463,7 @@ class BosonOnTorusShort :  public ParticleOnTorus
   // nbrBosons = number of bosons
   // maxMomentum = momentum maximum value for a boson
   // return value = Hilbert space dimension
-  int EvaluateHilbertSpaceDimension(int nbrBosons, int maxMomentum);
+  long EvaluateHilbertSpaceDimension(int nbrBosons, int maxMomentum);
 
   // evaluate Hilbert space dimension
   //
